4단계 - 자동차 경주(우승자) | refactor : InputView 클래스에서 입력 검증하는 기능을 분리
